+++ packages/create-twilio-function/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Ongoing [☰](https://github.com/twilio-labs/create-twilio-function/compare/v1.0.1...master)
 
-...
+- Minor updates
+  - Better error messages if the cli fails to create a directory. Fixes #14
 
 ## 1.0.1 (May 4, 2018) [☰](https://github.com/twilio-labs/create-twilio-function/compare/v1.0.0...v1.0.1)